Giver 상품 목록 및 상품 상태 변경 기능. 테스트 완료
