+++ src/main/java/com/cabbage/grabit/api/product/ProductApiController.java
@@ -2,12 +2,9 @@ package com.cabbage.grabit.api.product;
 
 import com.cabbage.grabit.config.aop.LogExecutionTime;
 import com.cabbage.grabit.domain.product.ProductRepository;
-import com.cabbage.grabit.domain.user.Giver;
 import com.cabbage.grabit.exception.ApiException;
 import com.cabbage.grabit.exception.ApiResult;
 import com.cabbage.grabit.exception.ApiStatus;
-import com.cabbage.grabit.domain.product.dto.request.ProductPostRequestDto;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,13 +17,6 @@ public class ProductApiController {
     private final ProductFacade productFacade;
     private final ProductRepository productRepository;
 
-    // TODO product 등록 시 이미지 파일 multipart 처리해야함
-    // TODO 인증 처리 해주어야 함
-    @PostMapping
-    public ApiResult save(@RequestBody ProductPostRequestDto requestDto){
-
-        return ApiResult.of(ApiStatus.SUCCESS, productFacade.postProduct(requestDto));
-    }
 
     @DeleteMapping("/{id}")
     public Long delete(@PathVariable Long id){
@@ -42,18 +32,6 @@ public class ProductApiController {
         return ApiResult.of(ApiStatus.SUCCESS, productService.selectProductListByGiver(giverId, page, size));
     }
 
-    @GetMapping("/giver/{productId}")
-    public ApiResult getMyProductDetail(@PathVariable Long productId,
-                                        @RequestParam Long giverId){
-
-        // FIXME 인증 코드로 바꿔야 함
-//        if(giver==null){
-//            throw new ApiException(ApiStatus.GIVER_NOT_FOUND);
-//        }
-
-        return ApiResult.of(ApiStatus.SUCCESS, productFacade.getProductDetailToGiver(productId, giverId));
-
-    }
 
     @LogExecutionTime
     @GetMapping
