+++ src/main/java/com/cabbage/grabit/api/product/ProductService.java
@@ -39,7 +39,7 @@ public class ProductService {
     @Transactional
     public Long postProduct(Giver giver, Set<Region> regionSet, ProductPostRequestDto requestDto){
         Product product = Product.create(giver, regionSet, requestDto);
-
+        System.out.println("happening?");
     return productRepository.save(product).getId();
     }
